+++ matlab/+urx/validate.m
@@ -1,4 +1,4 @@
-function res = saveToFile(dataset)
+function res = validate(dataset)
   libBinding = urx.LibBinding.getInstance();
   res = libBinding.call('urx_validate_dataset', dataset);
 end